WIP: investigating shader integer uploads

+++ shader/main_menu/vdialogue.shader
@@ -1,7 +1,7 @@
 #version 330 core
 
 in vec2 position;
-in float disp_id;
+in int disp_id;
 
 // opening transition
 uniform vec2 displace[3];
@@ -13,5 +13,5 @@ uniform mat4 proj;
 
 void main()
 {
-	gl_Position = proj*view*vec4(position+displace[int(disp_id)]*vec2(tprogress),.0,1.);
+	gl_Position = proj*view*vec4(position+displace[disp_id]*vec2(tprogress),.0,1.);
 }
